+++ apps/mock-oauth-server/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @stackframe/mock-oauth-server
 
+## 2.8.57
+
+## 2.8.56
+
 ## 2.8.55
 
 ## 2.8.54