+++ examples/js-example/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @stackframe/js-example
 
+## 2.8.57
+
+### Patch Changes
+
+- @stackframe/js@2.8.57
+
+## 2.8.56
+
+### Patch Changes
+
+- @stackframe/js@2.8.56
+
 ## 2.8.55
 
 ### Patch Changes