+++ packages/stack-shared/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @stackframe/stack-shared
 
+## 2.8.57
+
+### Patch Changes
+
+- Various changes
+
+## 2.8.56
+
+### Patch Changes
+
+- Various changes
+
 ## 2.8.55
 
 ### Patch Changes